Reliable communication between Python and ESP32
ESP32 code waits for Python script to send "#" character before it starts printing out lines (i.e. measurements) for the Python script to interpret.

Upon each restart of the ESP32 (which can be triggered using the physical EN button), the ESP32 sends a "@" character. The Python script detects this "@" character and restarts itself.